more tedlium updates

+++ users/hilmes/experiments/tedlium2/asr_2023/hybrid/wei_tedlium2_setup/experiment.py
@@ -13,7 +13,7 @@ def get_wei_config():
   nn_config = make_nn_config(network)
   nn_config['extern_data'] = {
     'data': {"dim": 80, "shape": (None, 80), "available_for_inference": True},  # input: 80-dimensional logmel features
-    'classes': {"dim": 9001, "shape": (None,), "available_for_inference": True, "sparse": True, "dtype": "int8"}
+    'classes': {"dim": 9001, "shape": (None,), "available_for_inference": True, "sparse": True, "dtype": "int16"}
   }
